rename: LinkOverlay.js to LinkOverlay.tsx
